Move text editor variant to separate component
